TODO : fix memory leak

+++ NeuralNetwork/NeuralNetwork/CInitializer.cpp
@@ -7,7 +7,7 @@ std::mt19937 CInitializer::mersenne = std::mt19937{ CInitializer::rd() };
 
 void CInitializer::NormalInitialize(CMatrix& mat)
 {
-	std::normal_distribution<float> dist{ 0.0, 1.0 };
+	std::normal_distribution<double> dist{ 0.0, 1.0 };
 	for (UINT idx = 0; idx < mat.size; ++idx)
 	{
 		mat.data[idx] = dist(CInitializer::mersenne);
@@ -16,7 +16,7 @@ void CInitializer::NormalInitialize(CMatrix& mat)
 
 void CInitializer::HeInitialize(CMatrix& mat, const UINT& numIn)
 {
-	std::normal_distribution<float> dist{ 0.0, std::sqrt(2.0f / static_cast<float>(numIn)) };
+	std::normal_distribution<double> dist{ 0.0, std::sqrt(2.0f / static_cast<double>(numIn)) };
 	for (UINT idx = 0; idx < mat.size; ++idx)
 	{
 		mat.data[idx] = dist(CInitializer::mersenne);
@@ -25,7 +25,7 @@ void CInitializer::HeInitialize(CMatrix& mat, const UINT& numIn)
 
 void CInitializer::XavierInitialize(CMatrix& mat, const UINT& numIn, const UINT& numOut)
 {
-	std::normal_distribution<float> dist{ 0.0, std::sqrt(2.0f / (static_cast<float>(numIn) + static_cast<float>(numOut))) };
+	std::normal_distribution<double> dist{ 0.0, std::sqrt(2.0f / (static_cast<double>(numIn) + static_cast<double>(numOut))) };
 	for (UINT idx = 0; idx < mat.size; ++idx)
 	{
 		mat.data[idx] = dist(CInitializer::mersenne);
